target_link_libraries: Properly link modules created by add_cxx_module
Setting one or more cxx_module to be linked to a binary turns on the
C++ Modules TS feature for the binary. Object file compilations must (at least
in Clang's case) see the module's object both at compilation and at
linking time.

This can be fine-tuned later for different compilers using the extra
variables introduced in this commit.